requirements.txt
Django>=5.2,<5.3
djangorestframework>=3.15,<4
gunicorn>=21.2,<24
whitenoise[brotli]>=6.6,<7
dj-database-url>=2.1,<4
psycopg2-binary>=2.9,<3
